Makefiles done

+++ libft/ft_calloc.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_calloc.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: djin <djin@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: lucas <lucas@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/30 19:58:29 by djin              #+#    #+#             */
-/*   Updated: 2023/05/09 18:40:22 by djin             ###   ########.fr       */
+/*   Updated: 2023/08/06 02:46:06 by lucas            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@ void	*ft_calloc(size_t num, size_t size)
 	size_t		i;
 
 	i = 0;
-	if (num == SIZE_MAX || size == SIZE_MAX)
+	if (num == INT_MAX || size == INT_MAX)
 		return (NULL);
 	ptr = (char *)malloc (num * size);
 	if (!ptr)
